% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{f_operator}
\alias{f_operator}
\title{Function operator for error handling}
\usage{
f_operator(f)
}
\arguments{
\item{f}{A function.}
}
\value{
A function that returns an error condition object when the input to f is invalid.
}
\description{
This function operator transforms a function so that it returns an error condition object instead of throwing an error when the input is invalid.
}
