% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{log.2}
\alias{log.2}
\title{Logarithm function}
\usage{
\method{log}{`2`}(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
The natural logarithm of x if x is positive. An error if x is non-positive.
}
\description{
This function computes the natural logarithm of a number and will throw an error if the input is non-positive.
}
