\name{package3-package}
\docType{package}
\title{
\packageTitle{package3}
}
\description{
\packageDescription{package3}
}
\details{

The DESCRIPTION file
\packageDESCRIPTION{package3}
\packageIndices{package3}
package contains implementations of two algorithm,Dijkstra and Euclidean
}
\author{ packageAuthor {Ashraful}

Maintainer packageMaintainer{Ashraful}
}

\keyword{GCD, Euclidean Algorithm }
\seealso{https://en.wikipedia.org/wiki/Greatest_common_divisor.
https://www.youtube.com/watch?v=gdmfOwyQlcI.}
\examples{euclidean(100, 1000)
dijkstra(wiki_graph, 1)
}
