% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pplot.R
\name{pplot}
\alias{pplot}
\title{A function that will plot the number of ARID1A mutations per cancer type}
\usage{
pplot(x)
}
\arguments{
\item{x}{- dataset to be used}
}
\value{
plot of number of ARID1A mutations per tumor type
}
\description{
This will dd two numbers, x and y
}
\examples{
pplot(x)
}

