% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{sqrt.2}
\alias{sqrt.2}
\title{Square root function}
\usage{
\method{sqrt}{`2`}(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
The square root of x if x is positive. An error if x is negative.
}
\description{
This function computes the square root of a number and throws an error if the input is negative.
}
