% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivial2.R
\name{trivial2}
\alias{trivial2}
\title{A Trivial function}
\usage{
trivial2(x, y)
}
\arguments{
\item{x, y}{two numbers to be multiplied}
}
\value{
the multiplication of two numbers
}
\description{
This will dd two numbers, x and y
}
\examples{
trivial2(3,2)
}

