% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package3.R
\docType{package}
\name{Package3}
\alias{Package3}
\alias{Package3-package}
\title{This is a description of Package 3}
\description{
I will probably include in this package mutation data for ARID1A that I'll get from TumorPortal
}
\details{
I don't know if I'll use it for anything else

Maybe I'll include some function in this package as well
}

